pyproject.toml
[tool.poetry]
name = "hymnal"
version = "0.1.0"
description = ""
authors = ["gersona <gersona@hotmail.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.3.2"
black = "^23.9.1"


[tool.poetry.group.dev.dependencies]
python-pptx = "^0.6.22"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
